% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestry_coefficients.r
\name{as.tfa_data}
\alias{as.tfa_data}
\title{Convert a data frame as an object of class tfa_data}
\usage{
as.tfa_data(dataframe)
}
\arguments{
\item{dataframe}{a dataframe containing metadata for a factor analysis of ancient DNA. The meta
data could contain information on sample IDs, Group.IDs, country of origin, sample age, etc. Columns
with names 'Instance.ID' and 'Group.ID' are mandatory.}
}
\value{
an object of class tfa_data
}
\description{
\code{as.tfa_data} checks the contents of a metadata file for the presence of group IDs, and
convert the data frame as an object of class tfa_data.
}
\examples{
library(tfa)

# Ancient DNA from Bronze Age Great Britain samples
# including Steppe (Yamnaya), hunter gatherers, and early farmers from Anatolia
data(england_ba)
attach(England_BA)
metadata <- as.tfa_data(meta)
detach(England_BA)
}
\references{
François, O., Jay, F. (2020). Factor analysis of ancient DNA samples.
}
\seealso{
\code{\link{England_BA}}, \code{\link{ancestry_coefficients}}
}
