% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestry_coefficients.r
\name{as.tfa_metadata}
\alias{as.tfa_metadata}
\title{Convert a data frame into an object of class \code{tfa_metadata}}
\usage{
as.tfa_metadata(dataframe)
}
\arguments{
\item{dataframe}{a dataframe containing metadata for a factor analysis of ancient DNA. The meta
data could contain information on individual ID's, group ID's, country of origin, sample age, etc. Columns
with names 'Instance.ID' and 'Group.ID' are mandatory.}
}
\value{
an object of class \code{tfa_metadata}
}
\description{
\code{as.tfa_metadata} checks the contents of a metadata file for the presence of group and instance IDs,
and it convert the data frame as an object of class \code{tfa_metadata}.
}
\examples{
library(tfa)

# Ancient DNA from Bronze Age Great Britain samples
# including Steppe (Yamnaya), hunter gatherers, and early farmers from Anatolia
data(england_ba)
attach(England_BA)
metadata <- as.tfa_metadata(meta)
detach(England_BA)
}
\references{
François, O., Jay, F. (2020). Factor analysis of ancient DNA samples.
}
\seealso{
\code{\link{England_BA}}, \code{\link{ancestry_coefficients}}
}
