% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/england_ba.R
\docType{data}
\name{England_BA}
\alias{England_BA}
\title{Ancient human DNA samples from David Reich's lab}
\format{List with 3 objects
\describe{
  \item{age}{ages, in years cal BP}
  \item{genotype}{a numeric matrix containing diploid genotypes (0,1,2) for
  118 ancient Eurasians. The SNPs were randomly selected from a much larger
  set of polymorphisms}
  \item{meta}{metadata on individuals: ID, group ID, country of origin and coverage}
}}
\source{
\url{https://reich.hms.harvard.edu/datasets}
}
\usage{
England_BA
}
\description{
A dataset containing 10,000 filtered SNP genotypes from 118 samples, from Neolithic and Bronze Age Great Britain,
Steppe (Yamnaya), early farmers from Anatolia, and hunter-gatherers from Serbia. Metadata and sample
ages are also provided.
}
\keyword{datasets}
