% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestry_coefficients.r
\name{ancestry_coefficients}
\alias{ancestry_coefficients}
\title{Compute ancestry coefficients from specified source populations}
\usage{
ancestry_coefficients(model, metadata, source, target, individual = FALSE)
}
\arguments{
\item{model}{an object of class \code{tfa} with the same number of individuals as in metadata.}

\item{metadata}{an object of class \code{tfa_metadata} containing metadata on Group.IDs.}

\item{source}{a vector of ancestral group ID's from the list of \code{Group.IDs} (character).}

\item{target}{a vector of target group ID's from the list of \code{Group.IDs} (character).}

\item{individual}{a logical indicating whether individual ancestry coefficients
should be computed.}
}
\value{
a matrix with coefficients equal to the ancestry proportions of each target population
from each source.
}
\description{
\code{ancestry_coefficients} compute the proportions of shared ancestry from specified source populations
}
\examples{
library(tfa)

# Ancient DNA from Bronze Age Great Britain samples
# including Steppe (Yamnaya), hunter gatherers, and early farmers from Anatolia
data(england_ba)
attach(England_BA)
metadata <- as.tfa_metadata(meta)
coverage <- meta$Coverage
geno <- coverage_adjust(genotype, coverage, K = 4, log = TRUE)

mod  <- tfa(age,
            geno,
            k = 3,
            lambda = 5e-1,
            center = TRUE,
            coverage = coverage,
            log = TRUE)

target = c("England_Bell_Beaker", "England_BA")
source = c("Anatolia_N", "Russia_Yamnaya", "Serbia_HG")
ancestry_coefficients(model = mod,
                      metadata = metadata,
                      source = source,
                      target = target)

source = c("Scotland_N", "Russia_Yamnaya", "Serbia_HG")
ancestry_coefficients(model = mod,
                      metadata = metadata,
                      source = source,
                      target = target)
detach(England_BA)
# rm(list = ls())
}
\references{
François, O., Jay, F. (2020). Factor analysis of ancient DNA samples. Under review.
}
\seealso{
\code{\link{England_BA}}, \code{\link{tfa}}, \code{\link{coverage_adjust}}, \code{\link{choose_lambda}}
}
\author{
Olivier Francois, \email{olivier.francois@univ-grenoble-alpes.fr}
}
