% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_adjust.R
\name{coverage_adjust}
\alias{coverage_adjust}
\title{Correct ancient genotypes for coverage effect}
\usage{
coverage_adjust(Y = NULL, coverage = NULL, K, lambda = 1e-05,
  log = FALSE)
}
\arguments{
\item{Y}{an nxp numeric data matrix containing genetic information for n individuals recorded in p columns.
Genetic information could be encoded as any numeric value, usually as genotypes taking values in 0,1,2.
Missing data are not allowed.}

\item{coverage}{a numerical vector containing information on each DNA sample coverage.}

\item{K}{a nonnegative integer for the number of components in the data matrix Y.
It could be obtained from a PCA of the data matrix.}

\item{lambda}{a ridge regularization parameter, to be kept at a small values (defaut 1e-5)).}

\item{log}{a logical value indicating that corrections are performed from log(coverage) instead of coverage.}
}
\value{
A list with the following attributes:
\describe{
   \item{W}{an nxp numeric matrix containing the n corrected genotypes}
}
}
\description{
\code{coverage_adjust} corrects the matrix of genotypes for low and uneven coverage of
individual genomes.
}
\examples{
library(tfa)

# Ancient DNA from Bronze Age Great Britain samples
# including Steppe (Yamnaya) and early farmers from Anatolia
data(england_ba)
attach(England_BA)
coverage <- as.numeric(as.character(meta$Coverage))
## TO DEVELOP
detach(England_BA)
}
\references{
François, O., Jay, F. (2020). Factor analysis of ancient DNA samples. Under review.
}
\seealso{
\code{\link{england_ba}}, \code{\link{tfa}}
}
