% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/england_ba.R
\docType{data}
\name{england_ba}
\alias{england_ba}
\title{Ancient human DNA samples from David Reich's lab}
\description{
{The data include 10,000 filtered SNP genotypes from 118 samples, from Neolithic and Bronze Age Great Britain,
Steppe (Yamnaya), early farmers from Anatolia, and hunter-gatherers from Serbia. The data set is structured as a list
named 'England_BA'

$age: vector of ages in  years cal BP.

$genotype: a numeric matrix containing 118 genotypes. The SNPs were randomly selected from a much larger set of variants.

$meta: metadata on individual ID, group ID, country of origin and coverage.

}
}
