% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/england_ba.R
\docType{data}
\name{england_ba}
\alias{england_ba}
\title{Ancient human DNA samples from David Reich's lab}
\description{
{The data include 143,081 filtered SNP genotypes from 64 samples, from Great Britain (Bronze Age), Yamnaya culture
and from early farmers from Anatolia (Olalde et al. Nature, 2018). The data set is structured as a list with name 'England_BA'

$age: vector of ages in  years cal BP.

$genotype: a large numeric matrix containing 64 genotypes corrected for coverage. The values,correspond to the residuals of a latent factor regression model with log(coverage)
considered as an explanatory variables.

$meta: metadata on individual ID, group ID, country of origin and coverage.

}
}
